+++ data/aircraft.json.d.ts
@@ -1,0 +1,13 @@
+// Type definitions for aircraft.json
+declare const AeroflyAircraft: Array<{
+  name: string;
+  nameFull: string;
+  icaoCode: string;
+  tags: string[];
+  approachAirspeedKts: number;
+  cruiseAltitudeFt: number;
+  cruiseSpeedKts: number;
+  maximumRangeNm: number;
+  aeroflyCode: string;
+}>;
+export default AeroflyAircraft;